+++ src/tests.rs
@@ -13,7 +13,7 @@ use arrayvec::ArrayVec;
 use core::cell::Cell;
 use core::mem::{align_of, uninitialized};
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 use super::{Box, Vec};
 
 // Struct that increments a counter each time it is dropped (used for testing
@@ -578,7 +578,7 @@ fn marker_extend_test() {
             .extend(allocation, [DropCounter::new(&drop_count)])
             .unwrap();
 
-        #[cfg(any(feature = "std", feature = "unstable"))]
+        #[cfg(any(feature = "std", feature = "alloc"))]
         let allocation = {
             let mut v = Vec::with_capacity(2);
             v.push(DropCounter::new(&drop_count));
@@ -604,7 +604,7 @@ fn marker_extend_test() {
         let _ = allocation; // Silence unused variable warnings...
     }
 
-    if cfg!(any(feature = "std", feature = "unstable")) {
+    if cfg!(any(feature = "std", feature = "alloc")) {
         assert_eq!(drop_count.get(), 8);
     } else {
         assert_eq!(drop_count.get(), 2);
@@ -668,7 +668,7 @@ fn slice_move_source_collection_tuple_test() {
 
         // `Box` and `Vec` may not be testable depending on the feature set,
         // so use different types if they're not available.
-        #[cfg(any(feature = "std", feature = "unstable"))]
+        #[cfg(any(feature = "std", feature = "alloc"))]
         let (b, c) = {
             let mut c = Vec::new();
             c.push(DropCounter::new(&drop_count));
@@ -684,7 +684,7 @@ fn slice_move_source_collection_tuple_test() {
             )
         };
 
-        #[cfg(not(any(feature = "std", feature = "unstable")))]
+        #[cfg(not(any(feature = "std", feature = "alloc")))]
         let (b, c) = (
             [DropCounter::new(&drop_count), DropCounter::new(&drop_count)],
             [
@@ -722,7 +722,7 @@ fn slice_move_source_collection_array_test() {
     assert_eq!(drop_count.get(), 3);
     drop_count.set(0);
 
-    #[cfg(any(feature = "std", feature = "unstable"))]
+    #[cfg(any(feature = "std", feature = "alloc"))]
     {
         {
             let marker = scratchpad.mark_back().unwrap();
@@ -759,7 +759,7 @@ fn slice_move_source_collection_array_test() {
 /// Tests vector implementations of `SliceMoveSourceCollection` use for
 /// possible leaks or unintentional dropping of elements.
 #[test]
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 fn slice_move_source_collection_vec_test() {
     let scratchpad = Scratchpad::<[usize; 3], [usize; 1]>::static_new();
     let drop_count = Cell::new(0);
@@ -827,7 +827,7 @@ fn slice_move_source_collection_vec_test() {
 /// Tests boxed slice implementations of `SliceMoveSourceCollection` use for
 /// possible leaks or unintentional dropping of elements.
 #[test]
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 fn slice_move_source_collection_boxed_slice_test() {
     let scratchpad = Scratchpad::<[usize; 3], [usize; 1]>::static_new();
     let drop_count = Cell::new(0);