+++ src/traits.rs
@@ -15,9 +15,9 @@ use core::str;
 use super::CacheAligned;
 use core::mem::{forget, size_of};
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 use super::{Box, Vec};
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 use core::mem::ManuallyDrop;
 
 /// Trait for types that can be safely used as the backing data type for
@@ -160,7 +160,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> Buffer for Box<[T]>
 where
     T: ByteData,
@@ -252,7 +252,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> SizeAlignedBuffer for Box<[T]>
 where
     T: SizeAlignedByteData,
@@ -682,7 +682,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> SliceSource<T> for Box<T>
 where
     T: SliceLike + ?Sized,
@@ -693,7 +693,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> SliceSource<[T]> for Vec<T> {
     #[inline]
     fn as_slice_like(&self) -> &[T] {
@@ -701,7 +701,7 @@ impl<T> SliceSource<[T]> for Vec<T> {
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'a, T> SliceSource<T> for &'a Box<T>
 where
     T: SliceLike + ?Sized,
@@ -712,7 +712,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'a, T> SliceSource<[T]> for &'a Vec<T> {
     #[inline]
     fn as_slice_like(&self) -> &[T] {
@@ -784,7 +784,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> SliceMoveSource<T> for Box<T>
 where
     T: SliceLike + ?Sized,
@@ -805,7 +805,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T> SliceMoveSource<[T]> for Vec<T> {
     fn move_elements<F>(self, mut f: F)
     where
@@ -817,7 +817,7 @@ impl<T> SliceMoveSource<[T]> for Vec<T> {
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'a, T> SliceMoveSource<T> for &'a Box<T>
 where
     T: SliceLike + ?Sized,
@@ -833,7 +833,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'a, T> SliceMoveSource<[T]> for &'a Vec<T>
 where
     T: Copy,
@@ -925,7 +925,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T, U> SliceSourceCollection<T> for Box<[U]>
 where
     T: SliceLike + ?Sized,
@@ -941,7 +941,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T, U> SliceSourceCollection<T> for Vec<U>
 where
     T: SliceLike + ?Sized,
@@ -957,7 +957,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'b, T, U> SliceSourceCollection<T> for &'b Box<[U]>
 where
     T: SliceLike + ?Sized,
@@ -973,7 +973,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'b, T, U> SliceSourceCollection<T> for &'b Vec<U>
 where
     T: SliceLike + ?Sized,
@@ -1057,7 +1057,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T, U> SliceMoveSourceCollection<T> for Box<[U]>
 where
     T: SliceLike + ?Sized,
@@ -1077,7 +1077,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<T, U> SliceMoveSourceCollection<T> for Vec<U>
 where
     T: SliceLike + ?Sized,
@@ -1093,7 +1093,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'b, T, U> SliceMoveSourceCollection<T> for &'b Box<[U]>
 where
     T: SliceLike + ?Sized,
@@ -1112,7 +1112,7 @@ where
     }
 }
 
-#[cfg(any(feature = "std", feature = "unstable"))]
+#[cfg(any(feature = "std", feature = "alloc"))]
 impl<'b, T, U> SliceMoveSourceCollection<T> for &'b Vec<U>
 where
     T: SliceLike + ?Sized,